refactor: extract bet sizer widget
